+++ packages/backend/package.json
@@ -15,6 +15,7 @@
     "ethers": "^4.0.45",
     "express": "^4.17.1",
     "knex": "^0.20.10",
+    "nodemon": "^2.0.2",
     "objection": "^2.1.3",
     "semaphore-auth-contracts": "file:../contracts",
     "sqlite3": "^4.1.1"
